Missed prototype for startTimeMs( unsigned cnt ) changed arg for DelayMs from long to unsigned int

+++ trunk/uc_code/interrupt.c
@@ -99,7 +99,7 @@ void startTimerMs( unsigned int cnt )
 	timerRunning = 1;
 }
 
-void DelayMs(unsigned long cnt)
+void DelayMs(unsigned cnt)
 {
 	startTimerMs( cnt );
 	while( timerRunning )continue;
